# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": { "title": "Collaborative Novel API", "version": "1.0.0" },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Error" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/stories": {
      "post": {
        "summary": "Create a new story",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "title": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "201": { "description": "Story created" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/stories/{id}": {
      "get": {
        "summary": "Get a story",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Story details" },
          "403": { "description": "Access denied" }
        }
      }
    },
    "/api/stories/{id}/invite": {
      "post": {
        "summary": "Invite a collaborator",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "role": { "type": "string", "enum": ["author", "editor", "commenter", "viewer"] },
                  "expiresAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Invite sent" },
          "403": { "description": "Access denied" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}